refactor: Cut out the Side_bar UI to a separate file
